display calculations for sum in adjustment for autocorrelation (mainly for debugging)

display calculations for sum in adjustment for autocorrelation (mainly for debugging)

sum_df(x, ...)

Arguments

x
numeric, a time series

...,
optional arguments for acf

Value

data frame showing terms in sum

Examples

set.seed(457298)
xx=stats::arima.sim(list(ar=0.8),100) # posoitively autocorrelated, sample size ratio > 1
sum_df(xx)

#>              v   n lags   sig      term     cumul   scaled
#> 1   1.00000000 100    0  TRUE 970200.00       0.0 1.000000
#> 2   0.78057306 100    1  TRUE 734592.62  734592.6 2.514312
#> 3   0.57306331 100    2  TRUE 522963.82 1257556.4 3.592365
#> 4   0.42732973 100    3  TRUE 378032.97 1635589.4 4.371654
#> 5   0.34517252 100    4  TRUE 295909.50 1931498.9 4.981651
#> 6   0.25211821 100    5  TRUE 209381.65 2140880.6 5.413277
#> 7   0.15402340 100    6 FALSE 123875.48 2140880.6 5.413277
#> 8   0.06866187 100    7 FALSE  53459.85 2140880.6 5.413277
#> 9  -0.02798080 100    8 FALSE -21082.97 2140880.6 5.413277
#> 10 -0.07326433 100    9 FALSE -53403.10 2140880.6 5.413277
#> 11 -0.07237324 100   10 FALSE -51014.45 2140880.6 5.413277
#> 12 -0.10307531 100   11 FALSE -70233.87 2140880.6 5.413277
#> 13 -0.14901890 100   12 FALSE -98116.43 2140880.6 5.413277
#> 14 -0.14209121 100   13 FALSE -90365.75 2140880.6 5.413277
#> 15 -0.12443444 100   14 FALSE -76407.73 2140880.6 5.413277
#> 16 -0.09222622 100   15 FALSE -54655.10 2140880.6 5.413277
#> 17 -0.07299130 100   16 FALSE -41729.42 2140880.6 5.413277
#> 18 -0.07996700 100   17 FALSE -44084.69 2140880.6 5.413277
#> 19 -0.09675968 100   18 FALSE -51414.22 2140880.6 5.413277
#> 20 -0.12888989 100   19 FALSE -65981.31 2140880.6 5.413277
#> 21 -0.19530753 100   20 FALSE -96278.80 2140880.6 5.413277